# backend/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Alpine Ecommerce API",
    "version": "1.0.0",
    "description": "# Alpine Ecommerce Backend API\n\nComplete API documentation for Alpine Ecommerce e-commerce platform.\n\n## Features\n- 🔐 JWT Authentication\n- 📦 Product Management\n- 🛒 Order Processing\n- 💳 Payment Integration (Razorpay)\n- 👥 Customer Management\n- 📊 Dashboard & Reports\n- 📝 Error Logging\n\n## Authentication\nMost endpoints require JWT authentication. Include the token in the Authorization header:\n```\nAuthorization: Bearer <your-token>\n```\n\n## Base URL\n- **Current Server**: http://13.235.53.15:8081/api/v1\n- **Configuration**: \n  - Set `API_BASE_URL` in .env for full control (recommended)\n  - Or set `PRODUCTION_URL` for production domain\n  - Or set `DOMAIN`, `PORT`, and `PROTOCOL` for auto-construction\n\n## Response Format\nAll responses follow a consistent format:\n```json\n{\n  \"responseStatusCode\": 200,\n  \"isError\": false,\n  \"errorData\": null,\n  \"successData\": { ... }\n}\n```\n\n## Error Handling\nErrors are returned with appropriate status codes:\n```json\n{\n  \"responseStatusCode\": 400,\n  \"isError\": true,\n  \"errorData\": {\n    \"displayMessage\": \"Error description\",\n    \"apiErrorType\": 1\n  }\n}\n```",
    "contact": {
      "name": "Alpine Ecommerce API Support",
      "email": "support@wildvalleyfoods.in"
    },
    "license": {
      "name": "Proprietary"
    },
    "servers": [
      {
...
      }
    ],
...
        }
...
  },
  "paths": {
...
    "/api/v1/product/search": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "Search products (optimized)",
        "description": "Search products by name. Returns only essential data (id, name, thumbnail) for fast autocomplete/suggestions.\n- Returns only active products (isActive = true)\n- Returns only first image per product as compressed WebP thumbnail (200x200)\n- Default limit is 6 results\n- Supports both 'query' and 'q' parameters (backward compatible)\n",
        "security": [],
        "parameters": [
          {
            "in": "query",
            "name": "query",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Search keyword (minimum 1 character). Also accepts 'q' parameter for backward compatibility.",
            "example": "apple"
          },
          {
            "in": "query",
            "name": "q",
            "schema": {
              "type": "string"
            },
            "required": false,
            "description": "Alternative search keyword parameter (backward compatible)",
            "example": "banana"
          },
          {
            "in": "query",
            "name": "top",
            "schema": {
              "type": "integer",
              "default": 6,
              "minimum": 1,
              "maximum": 50
            },
            "description": "Maximum number of results to return (default 6).",
            "example": 10
          },
          {
            "in": "query",
            "name": "skip",
            "schema": {
              "type": "integer",
              "default": 0,
              "minimum": 0
            },
            "description": "Number of results to skip for pagination.",
            "example": 0
          }
        ],
        "responses": {
          "200": {
            "description": "Products retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "responseStatusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "isError": {
                      "type": "boolean",
                      "example": false
                    },
                    "responseData": {
                      "type": "object",
                      "properties": {
                        "products": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "integer",
                                "example": 1
                              },
                              "name": {
                                "type": "string",
                                "example": "Apple"
                              },
                              "thumbnail": {
                                "type": "string",
                                "nullable": true,
                                "description": "Base64 encoded WebP thumbnail (200x200) or null if no image",
                                "example": "data:image/webp;base64,UklGRiQAAABXRUJQVlA4IBgAAAAwAQCdASoCAAEAAkA4JQBdAD6jQAA3p..."
                              }
                            }
                          }
                        },
                        "total": {
                          "type": "integer",
                          "description": "Total number of matching products",
                          "example": 25
                        },
                        "count": {
                          "type": "integer",
                          "description": "Number of products in current response",
                          "example": 6
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - query parameter missing or too short",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "responseStatusCode": {
                      "type": "integer",
                      "example": 400
                    },
                    "isError": {
                      "type": "boolean",
                      "example": true
                    },
                    "errorData": {
                      "type": "object",
                      "properties": {
                        "displayMessage": {
                          "type": "string",
                          "example": "Please enter at least 1 character"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No products found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "responseStatusCode": {
                      "type": "integer",
                      "example": 404
                    },
                    "isError": {
                      "type": "boolean",
                      "example": true
                    },
                    "errorData": {
                      "type": "object",
                      "properties": {
                        "displayMessage": {
                          "type": "string",
                          "example": "No products found for this keyword"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/product/paginated": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "Get paginated products",
        "description": "Returns products with pagination and OData support",
        "security": [],
        "parameters": [
          {
            "in": "query",
            "name": "$filter",
            "schema": {
              "type": "string"
            },
            "description": "OData filter expression"
          },
          {
            "in": "query",
            "name": "$orderby",
            "schema": {
              "type": "string"
            },
            "description": "OData orderby expression"
          },
          {
            "in": "query",
            "name": "$skip",
            "schema": {
              "type": "integer"
            },
            "description": "Number of records to skip"
          },
          {
            "in": "query",
            "name": "$top",
            "schema": {
              "type": "integer"
            },
            "description": "Number of records to return"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated products retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/product/ByCategoryId/{categoryId}/paginated": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "Get paginated products by category",
        "description": "Returns products filtered by category ID with pagination",
        "security": [],
        "parameters": [
          {
            "in": "path",
            "name": "categoryId",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Category ID"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "Page number"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "Number of items per page"
          }
        ],
        "responses": {
          "200": {
            "description": "Products retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/product/count/ByCategoryId/{categoryId}": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "Get product count by category",
        "description": "Returns the total number of products in a category",
        "security": [],
        "parameters": [
          {
            "in": "path",
            "name": "categoryId",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Category ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Product count retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/product/{id}": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "Get product by ID",
        "description": "Returns a specific product by its ID",
        "security": [],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Product ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Product retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from /register or /login endpoint. Enter your token (with or without \"Bearer \" prefix)"
      }
    }
  },
...
}